Change GetAlerts Handler to a separate file
